# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cafe Supply Chain API",
        "description": "REST API documentation split by platform: Admin Dashboard, Cafe Mobile App, and Public Storefront.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/delegates": {
            "get": {
                "tags": [
                    "Delegates"
                ],
                "summary": "List delegates",
                "operationId": "18166f91a2923c9cc5343e78e17a4391",
                "responses": {
                    "200": {
                        "description": "Paginated list of delegates"
                    }
                }
            },
            "post": {
                "tags": [
                    "Delegates"
                ],
                "summary": "Create a delegate",
                "operationId": "ca759c5a311eb221fd8a470e0ea8858d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DelegateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Delegate created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/delegates/{id}": {
            "get": {
                "tags": [
                    "Delegates"
                ],
                "summary": "Get a delegate",
                "operationId": "917c6661df9cd3a085af193a88390625",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delegate details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Delegates"
                ],
                "summary": "Update a delegate",
                "operationId": "c8689bb8ef2a69700d4153ae156a550f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DelegateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Delegate updated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Delegates"
                ],
                "summary": "Delete a delegate",
                "operationId": "f8f0b98647e9adb64bcc04806802ee9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delegate deleted"
                    }
                }
            }
        },
        "/delegates/{id}/toggle-active": {
            "post": {
                "tags": [
                    "Delegates"
                ],
                "summary": "Toggle delegate active status",
                "operationId": "75427988c2ceb540a7b89a0b2bdfd7b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status toggled"
                    }
                }
            }
        },
        "/delegates/{id}/location": {
            "put": {
                "tags": [
                    "Delegates"
                ],
                "summary": "Update delegate location from admin",
                "operationId": "09985dd6ce0c05ee65f34f97227d17ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "is_available": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated"
                    }
                }
            }
        },
        "/delegate/location": {
            "post": {
                "tags": [
                    "Delegate Mobile"
                ],
                "summary": "Update delegate live location",
                "operationId": "db1599a7f2455579d7c4006c83de8320",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DelegateLocationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/delegate/availability": {
            "post": {
                "tags": [
                    "Delegate Mobile"
                ],
                "summary": "Toggle delegate availability",
                "operationId": "ee0587636fba3123435e4ea7d1edc612",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DelegateAvailabilityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Availability updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/delegate/orders": {
            "get": {
                "tags": [
                    "Delegate Mobile"
                ],
                "summary": "List assigned orders for the delegate",
                "operationId": "e5d9c58241b4f14bad1ef8f4e3f738cd",
                "responses": {
                    "200": {
                        "description": "List of orders"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/delegate/orders/{id}": {
            "get": {
                "tags": [
                    "Delegate Mobile"
                ],
                "summary": "Get assigned order details",
                "operationId": "1cc0f4360c07606062d21188acc2c3f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/delegate/orders/{id}/status": {
            "post": {
                "tags": [
                    "Delegate Mobile"
                ],
                "summary": "Update status of an assigned order",
                "operationId": "45ff765c6dc091e5c2f6db889aac118b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DelegateRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 150
                    },
                    "mobile_number": {
                        "type": "string",
                        "maxLength": 20,
                        "nullable": true
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minLength": 6
                    },
                    "cafe_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "DelegateLocationRequest": {
                "required": [
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "maximum": 90,
                        "minimum": -90
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "maximum": 180,
                        "minimum": -180
                    }
                },
                "type": "object"
            },
            "DelegateAvailabilityRequest": {
                "required": [
                    "is_available"
                ],
                "properties": {
                    "is_available": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}